Time: 1 ms (99.96%) | Memory: 42.7 MB (51.27%) - LeetSync

+++ 4135-concatenate-non-zero-digits-and-multiply-by-sum-i/concatenate-non-zero-digits-and-multiply-by-sum-i.java
@@ -2,7 +2,7 @@ class Solution {
     public long sumAndMultiply(int n) {
         long x = 0;
         long sum = 0;
-        long len= 1;
+        int len= 1;
         while(n!=0){
             int digit = n%10;
